refactor(workflows): drop the dead duplicate workflows list modules
Three files under modules/ee/workflows/list/ imported only each other and
nothing outside the directory imported them. Each was a superseded copy of a
live module the workflows list page actually uses:

  list/hooks/use-workflows.ts      -> hooks/use-workflows.ts
  list/lib/v3-workflows-client.ts  -> lib/api-client.ts
  list/lib/query.ts                -> lib/query.ts

The dead useWorkflows took no filter arguments at all — no nameContains,
statusIn or sortBy — so it predates the list filtering the live hook carries,
and its workflowKeys.list() keyed on { workspaceId, limit } alone.

use-trigger-survey-picker.ts is the one live module in that directory, so it
moves up to hooks/ (with its test) and list/ goes away entirely. The hook
imports nothing relative, so the move needed no edits inside it — only the
four import sites.
